% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat.R
\name{as.feat}
\alias{as.feat}
\title{Convert objects to feat class}
\usage{
as.feat(x)
}
\arguments{
\item{x}{Object to convert (character, list, or feat).}
}
\value{
A feat object.
}
\description{
Convert objects to feat class
}
