% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typMisc.R
\name{as.typstcode}
\alias{as.typstcode}
\title{Create a typstcode object}
\usage{
as.typstcode(x)
}
\arguments{
\item{x}{An object to be coerced to class typstcode}
}
\value{
The object with class typstcode prepended
}
\description{
Create a typstcode object
}
