% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typMisc.R
\name{breakable}
\alias{breakable}
\title{Set figure breakability in Typst}
\usage{
breakable(x = TRUE)
}
\arguments{
\item{x}{Logical indicating if figures should be breakable}
}
\value{
A typstcode object containing the Typst show rule
}
\description{
Set figure breakability in Typst
}
