% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typMisc.R
\name{emit_typstcode}
\alias{emit_typstcode}
\title{Format string as a Typst raw block}
\usage{
emit_typstcode(x)
}
\arguments{
\item{x}{A character string}
}
\value{
A string wrapped in Quarto's Typst raw attribute fences
}
\description{
Format string as a Typst raw block
}
