% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emph.R
\name{encircle}
\alias{encircle}
\title{draw circle around a string}
\usage{
encircle(x, radius = "8pt", stroke = "1pt", color = "black")
}
\arguments{
\item{x}{character scalar/vector.}

\item{radius}{string defining adius of circule surrounding x}

\item{stroke}{string defining thickness of circle}

\item{color}{string defining color of circle}
}
\description{
draw circle around a string
}
