% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat.R
\name{feat}
\alias{feat}
\title{Create a feat object}
\usage{
feat(pre = "", epi = "")
}
\arguments{
\item{pre}{Preamble string.}

\item{epi}{Epilogue string.}
}
\value{
An S3 object of class 'feat'.
}
\description{
Create a feat object
}
