% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat.R
\name{feat_align}
\alias{feat_align}
\title{Generate a Typst layout environment (Alignment & Blocks)}
\usage{
feat_align(align = "center", inner_align = "left")
}
\arguments{
\item{align}{String. Outer alignment of the block (e.g., "center").}

\item{inner_align}{String. Alignment inside the block (e.g., "left").}
}
\description{
Generate a Typst layout environment (Alignment & Blocks)
}
