% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat.R
\name{feat_align_cells}
\alias{feat_align_cells}
\title{Align all cells in table}
\usage{
feat_align_cells(align = c("left", "center", "right"))
}
\arguments{
\item{align}{Character. One of "left", "center", or "right".}
}
\value{
A feat object.
}
\description{
Align all cells in table
}
