% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat.R
\name{feat_colspan_font_size}
\alias{feat_colspan_font_size}
\title{Style spanned table cells}
\usage{
feat_colspan_font_size(num, size = "8pt")
}
\arguments{
\item{num}{Integer. The colspan value to target.}

\item{size}{String. The font size for these specific cells.}
}
\value{
A feat object.
}
\description{
Style spanned table cells
}
