% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat.R
\name{feat_notes}
\alias{feat_notes}
\title{Generate Typst table notes and environment}
\usage{
feat_notes(
  notes,
  notes_size = "0.8em",
  align = "center",
  table_font_size = "8pt",
  vert = "-1em"
)
}
\arguments{
\item{notes}{String. The text of the note to appear below the table.}

\item{notes_size}{String. Font size of the note (e.g., "0.8em").}

\item{align}{String. Outer alignment of the table on the page.}

\item{table_font_size}{String. Font size for the table body.}

\item{vert}{String. Vertical adjustment to "glue" the note to the table.}
}
\value{
A feat object.
}
\description{
Generate Typst table notes and environment
}
