% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typMisc.R
\name{knit_print.typstcode}
\alias{knit_print.typstcode}
\title{knit_print method for typstcode objects}
\usage{
knit_print.typstcode(x, ...)
}
\arguments{
\item{x}{A typstcode object}

\item{...}{Unused}
}
\description{
knit_print method for typstcode objects
}
