% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat.R
\name{+.feat}
\alias{+.feat}
\title{Combine Typst features}
\usage{
\method{+}{feat}(e1, e2)
}
\arguments{
\item{e1}{A feat object or character string.}

\item{e2}{A feat object or character string.}
}
\description{
Combines features. Preambles are concatenated in order.
Epilogues are concatenated in reverse (LIFO) to ensure nested blocks close correctly.
}
