% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat.R
\name{print.feat}
\alias{print.feat}
\title{Print method for feat objects}
\usage{
\method{print}{feat}(x, ...)
}
\arguments{
\item{x}{A feat object.}

\item{...}{Unused.}
}
\description{
Print method for feat objects
}
