% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typMisc.R
\name{print.typstcode}
\alias{print.typstcode}
\title{Print method for typstcode objects}
\usage{
\method{print}{typstcode}(x, ...)
}
\arguments{
\item{x}{A typstcode object}

\item{...}{Unused}
}
\value{
The object x, invisibly
}
\description{
Print method for typstcode objects
}
