% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{render_rtyp}
\alias{render_rtyp}
\title{render .Rtyp file processing it with knitr and then typst
Must install binary from https://github.com/typst/typst/releases
or build from rust source: cargo install --locked typst-cli}
\usage{
render_rtyp(x, ...)
}
\description{
Assumes typst is on PATH.
param x input file.  Use path if not in current directory.
... not used
}
