% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typMisc.R
\name{table_note}
\alias{table_note}
\title{Wrap a table with a footer note in Typst}
\usage{
table_note(
  ht,
  notes = "",
  size = "0.8em",
  vert = "-1em",
  align = c("center", "left", "right")
)
}
\arguments{
\item{ht}{A huxtable object}

\item{notes}{Character string or vector of notes}

\item{size}{Typst text size for the note}

\item{vert}{Vertical spacing before the note}

\item{align}{Alignment of the block: "center", "left" or "right"}
}
\value{
A typstcode object
}
\description{
Wrap a table with a footer note in Typst
}
