% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat.R
\name{tt_env}
\alias{tt_env}
\title{Wrapper for typstable::tt with environment support}
\usage{
tt_env(x, ..., env = NULL, preamble = NULL, epilogue = NULL)
}
\arguments{
\item{x}{A data frame.}

\item{...}{Additional arguments passed to typstable::tt.}

\item{env}{A feat object (or combined feat objects using +).}

\item{preamble}{Optional character string (ignored if env is provided).}

\item{epilogue}{Optional character string (ignored if env is provided).}
}
\description{
Wrapper for typstable::tt with environment support
}
