% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat.R
\name{feat_notes}
\alias{feat_notes}
\title{Generate Typst table notes}
\usage{
feat_notes(
  notes,
  notes_size = "0.8em",
  align = "center",
  table_font_size = "8pt",
  vert = "-1em"
)
}
\arguments{
\item{align}{String. Alignment of table on page.}

\item{vert}{String. Vertical adjustment to "glue" the note to the table (e.g., "-1em").}

\item{notes.}{String giving text of notes.}

\item{notes_size.}{Font size of notes.}

\item{text}{String. The text of the note to appear below the table.}

\item{size}{String. Font size of the note (e.g., "0.8em").}
}
\value{
A list containing the epilogue string.
}
\description{
Generate Typst table notes
}
