% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat.R
\name{feat_text_size}
\alias{feat_text_size}
\title{Set table font size}
\usage{
feat_text_size(size = "8pt")
}
\arguments{
\item{size}{String. The font size (e.g., "9pt").}
}
\value{
A feat object.
}
\description{
Set table font size
}
